Move pages to their own directory, cleaned up imports in index while I was at it
