+++ app/app/pages/Movies/components/Card/index.js
