+++ app/app/pages/Movies/components/Card/shape.js
